# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * SecureDMS prototype API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface LoginInput {
  email: string;
  password: string;
}

export interface User {
  id: string;
  name: string;
  email: string;
  role: string;
  initials: string;
  active: boolean;
}

export interface AuthSession {
  user: User;
  prototype: boolean;
}

export interface CaseInput {
  title: string;
  category: string;
  station: string;
}

export interface Case {
  id: string;
  title: string;
  category: string;
  station: string;
  officer: string;
  opened: string;
  status: string;
  documents: number;
  activity: string;
}

export type CaseDetailsDatesItem = {
  label: string;
  value: string;
};

export type CaseDetails = Case & {
  description: string;
  assigned: string[];
  dates: CaseDetailsDatesItem[];
};

export interface DocumentInput {
  caseId: string;
  title: string;
  type: string;
  fileName: string;
  description?: string;
  confidentiality?: string;
}

export interface Document {
  id: string;
  caseId: string;
  title: string;
  type: string;
  fileName: string;
  version: string;
  uploadedBy: string;
  date: string;
  integrity: string;
  signature: string;
  hash: string;
  classification: string;
  confidence: number;
  size: string;
  confidentiality?: string;
}

export interface DocumentVersion {
  version: string;
  createdBy: string;
  timestamp: string;
  hash: string;
  change: string;
  integrity: string;
}

export type DocumentDetail = Document & {
  ocrText: string;
  originalHash: string;
  currentHash: string;
  signedBy: string;
  versions: DocumentVersion[];
};

export interface IntegrityResult {
  documentId: string;
  status: string;
  originalHash: string;
  currentHash: string;
  message: string;
}

export interface AuditLog {
  id: string;
  timestamp: string;
  user: string;
  role: string;
  action: string;
  caseId: string;
  document: string;
  device: string;
  status: string;
}

export interface SearchResult {
  document: Document;
  score: number;
  match: string;
}

export type DashboardMetrics = {
  activeCases: number;
  totalDocuments: number;
  pendingReviews: number;
  verified: number;
  alerts: number;
  addedToday: number;
};

export type DashboardUploadsItem = {
  day: string;
  count: number;
};

export type DashboardSecurity = {
  score: number;
  signed: number;
  encrypted: number;
};

export interface Dashboard {
  metrics: DashboardMetrics;
  uploads: DashboardUploadsItem[];
  activity: AuditLog[];
  security: DashboardSecurity;
}

export type CaseIdParameter = string;

export type SearchParameter = string;

export type StatusParameter = string;

export type GetCasesParams = {
status?: StatusParameter;
q?: SearchParameter;
};

export type GetDocumentsParams = {
caseId?: CaseIdParameter;
q?: SearchParameter;
};

export type SearchDocumentsParams = {
q?: SearchParameter;
};