# lib/api-zod/src/generated/types/authSession.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * SecureDMS prototype API
 * OpenAPI spec version: 0.1.0
 */
import type { User } from './user';

export interface AuthSession {
  user: User;
  prototype: boolean;
}