# lib/api-zod/src/generated/types/case.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * SecureDMS prototype API
 * OpenAPI spec version: 0.1.0
 */

export interface Case {
  id: string;
  title: string;
  category: string;
  station: string;
  officer: string;
  opened: string;
  status: string;
  documents: number;
  activity: string;
}