# lib/api-zod/src/generated/types/caseDetails.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * SecureDMS prototype API
 * OpenAPI spec version: 0.1.0
 */
import type { Case } from './case';
import type { CaseDetailsDatesItem } from './caseDetailsDatesItem';

export type CaseDetails = Case & {
  description: string;
  assigned: string[];
  dates: CaseDetailsDatesItem[];
};