# lib/api-zod/src/generated/types/caseInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * SecureDMS prototype API
 * OpenAPI spec version: 0.1.0
 */

export interface CaseInput {
  title: string;
  category: string;
  station: string;
}