# lib/api-zod/src/generated/types/dashboard.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * SecureDMS prototype API
 * OpenAPI spec version: 0.1.0
 */
import type { AuditLog } from './auditLog';
import type { DashboardMetrics } from './dashboardMetrics';
import type { DashboardSecurity } from './dashboardSecurity';
import type { DashboardUploadsItem } from './dashboardUploadsItem';

export interface Dashboard {
  metrics: DashboardMetrics;
  uploads: DashboardUploadsItem[];
  activity: AuditLog[];
  security: DashboardSecurity;
}