# lib/api-zod/src/generated/types/document.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * SecureDMS prototype API
 * OpenAPI spec version: 0.1.0
 */

export interface Document {
  id: string;
  caseId: string;
  title: string;
  type: string;
  fileName: string;
  version: string;
  uploadedBy: string;
  date: string;
  integrity: string;
  signature: string;
  hash: string;
  classification: string;
  confidence: number;
  size: string;
  confidentiality?: string;
}