# lib/api-zod/src/generated/types/documentDetail.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * SecureDMS prototype API
 * OpenAPI spec version: 0.1.0
 */
import type { Document } from './document';
import type { DocumentVersion } from './documentVersion';

export type DocumentDetail = Document & {
  ocrText: string;
  originalHash: string;
  currentHash: string;
  signedBy: string;
  versions: DocumentVersion[];
};