# lib/api-zod/src/generated/types/documentInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * SecureDMS prototype API
 * OpenAPI spec version: 0.1.0
 */

export interface DocumentInput {
  caseId: string;
  title: string;
  type: string;
  fileName: string;
  description?: string;
  confidentiality?: string;
}