# lib/api-zod/src/generated/types/documentVersion.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * SecureDMS prototype API
 * OpenAPI spec version: 0.1.0
 */

export interface DocumentVersion {
  version: string;
  createdBy: string;
  timestamp: string;
  hash: string;
  change: string;
  integrity: string;
}