# lib/api-zod/src/generated/types/integrityResult.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * SecureDMS prototype API
 * OpenAPI spec version: 0.1.0
 */

export interface IntegrityResult {
  documentId: string;
  status: string;
  originalHash: string;
  currentHash: string;
  message: string;
}