# lib/api-zod/src/generated/types/searchResult.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * SecureDMS prototype API
 * OpenAPI spec version: 0.1.0
 */
import type { Document } from './document';

export interface SearchResult {
  document: Document;
  score: number;
  match: string;
}